Base,BaseGL: clang tidy

+++ modules/base/src/processors/volumesliceextractor.cpp
@@ -339,7 +339,8 @@ std::shared_ptr<Image> extractSlice(const VolumeRAMPrecision<T>* vrprecision,
 void VolumeSliceExtractor::process() {
     auto vol = inport_.getData();
 
-    if (tfGroup_.isChecked() && channel_ >= vol->getDataFormat()->getComponents()) {
+    if (tfGroup_.isChecked() &&
+        channel_ >= static_cast<int>(vol->getDataFormat()->getComponents())) {
         throw Exception(SourceContext{},
                         "Channel for TF mapping is greater than the available channels {} >= {}",
                         channel_.get() + 1, vol->getDataFormat()->getComponents());
